backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
elasticsearch==8.11.1
python-dotenv
